Cleanup
Cleaning up the code for creating the list of products.

Instead of a whole heck of a lot of if statements, I've made a reusable function to create the HTML. It accepts the container and array as arguments and creates HTML in the container based on the array.

The array is determined by whether or not the checkbox is checked