Made a reusable function for the list of products
Made a reusable function for the list of products, meaning I don't have to manage two separate functions to create lists of products.

The function takes a container and an array as arguments, then makes HTML dynamically based on that.

For this to work I have to filter my arrays before running the function. Seems to work well